+++ csrc/include/natten/cuda/fmha_hopper/kernel/fmha_kernel_builder.hpp
@@ -65,9 +65,9 @@ struct FmhaBuilder<
     ElementAccumulator,
     ElementAccumulator,
     TileShape,
-    cute::tuple<int, _1, cute::tuple<int, int>>,
-    cute::tuple<int, _1, cute::tuple<int, int>>,
-    cute::tuple<int, _1, cute::tuple<int, int>>,
+    cute::tuple<int, _1, cute::tuple<int64_t, int>>,
+    cute::tuple<int, _1, cute::tuple<int64_t, int>>,
+    cute::tuple<int, _1, cute::tuple<int64_t, int>>,
     Fusion,
     cutlass::gemm::KernelTma,
     Options...> {
